Feat/qh dashboard

+++ src/main/java/com/donut/prokindonutsweb/dashboard/controller/QhDashboardController.java
@@ -51,7 +51,7 @@ public class QhDashboardController{
         OrderInboundDTO dto = qhDashboardService.getOrderInbound(warehouseCode);
         //입출고 진행량 객체 반환
         return new OrderInboundRateDTO(
-                dto.inboundApprovalRate() * 100, dto.orderedRate() * 100);
+                (int)dto.inboundApprovalRate() * 100, (int)dto.orderedRate() * 100);
     }
 
 }
